Remove iojs support

+++ js.sh
@@ -4,13 +4,8 @@ set -e
 
 # [[ -f .jsshrc ]] && eval $(cat .jsshrc)
 
-: ${NODE_ENV:=development}
-
 : ${NODE_DIST:=node-v4.0.0}
-# : ${NODE_DIST:=iojs-v3.3.0}
-# : ${NODE_DIST:=node-v0.10.40}
-# : ${NODE_DIST:=node-v0.12.7}
-# : ${NODE_DIST:=iojs-v2.4.1-nightly2015072079c865a53f}
+: ${NODE_ENV:=development}
 
 NODE_TYPE=${NODE_DIST%%-*}
 NODE_VER=${NODE_DIST#*-}
@@ -18,7 +13,7 @@ NODE_OS=$(uname | tr A-Z a-z)
 NODE_DIR="vendor/$NODE_TYPE-$NODE_VER-$NODE_OS-x64"
 
 if [[ $# == 0 ]]; then
-  echo 'usage: js.sh [--node-bin] [--iojs-bin] [--npm-bin] [--clean] [--env]'
+  echo 'usage: js.sh [--node-bin] [--npm-bin] [--clean] [--env]'
   echo '       js.sh [--clean] [node|npm|MODULE_CLI] [args...]'
   exit 1
 fi
@@ -26,12 +21,12 @@ fi
 while test $# -gt 0; do
   case "$1" in
     --node-bin) echo $NODE_DIR/bin/node; shift;;
-    --iojs-bin) echo $NODE_DIR/bin/iojs; shift;;
     --npm-bin)  echo $NODE_DIR/bin/npm; shift;;
     --clean)
-      [[ -d "vendor" ]] && \
-        find vendor -maxdepth 1 \( -type d -name "iojs-v*" -o -name "node-v*" \) \
-        -exec sh -c 'echo "js.sh: Removing {}" 1>&2; rm -rf "{}"' \;
+      if [[ -d 'vendor' ]]; then
+        find 'vendor' -maxdepth 1 -type d -name 'node-v*' \
+          -exec sh -c 'echo "js.sh: Removing {}" 1>&2; rm -rf "{}"' \;
+      fi
       shift
       ;;
     --env)
@@ -51,13 +46,7 @@ done
 [[ $# == 0 ]] && exit 0
 
 if [[ ! -e $NODE_DIR/bin/node ]] || [[ ! -e $NODE_DIR/bin/npm ]]; then
-  if [[ $NODE_TYPE == iojs ]] && [[ $NODE_VER == *"nightly"* ]]; then
-    NODE_URL="https://iojs.org/download/nightly/$NODE_VER/iojs-$NODE_VER-$NODE_OS-x64.tar.gz"
-  elif [[ $NODE_TYPE == iojs ]]; then
-    NODE_URL="https://iojs.org/dist/$NODE_VER/iojs-$NODE_VER-$NODE_OS-x64.tar.gz"
-  elif [[ $NODE_TYPE == node ]]; then
-    NODE_URL="https://nodejs.org/dist/$NODE_VER/node-$NODE_VER-$NODE_OS-x64.tar.gz"
-  fi
+  NODE_URL="https://nodejs.org/dist/$NODE_VER/$NODE_TYPE-$NODE_VER-$NODE_OS-x64.tar.gz"
   echo "js.sh: Downloading $NODE_URL ..." 1>&2
   mkdir -p $NODE_DIR
   curl $NODE_URL | tar -xz -C $NODE_DIR --strip-components=1
@@ -65,7 +54,7 @@ fi
 
 export PATH="$PWD/$NODE_DIR/bin:$PATH"
 export NODE_PATH="$PWD/$NODE_DIR/lib/node_modules"
-export NODE_ENV=$NODE_ENV
+export NODE_ENV="$NODE_ENV"
 
 NODE_CMD_BIN="$NODE_DIR/bin/$1"
 MODULE_CLI_BIN="node_modules/.bin/$1"
